fix: support public for batch_key in Transaction
