+++ tests/unit/batch_transaction_test.py
@@ -479,3 +479,47 @@ def test_batch_key_mixed_private_and_public_keys(mock_client, mock_account_ids):
     assert len(batch_tx.inner_transactions) == 2
     assert isinstance(batch_tx.inner_transactions[0].batch_key, PrivateKey)
     assert isinstance(batch_tx.inner_transactions[1].batch_key, type(public_key))
+
+
+def test_set_batch_key_with_private_key():
+    """Test that batch_key can be set with PrivateKey."""
+    private_key = PrivateKey.generate_ed25519()
+    transaction = TransferTransaction()
+    
+    result = transaction.set_batch_key(private_key)
+    
+    assert transaction.batch_key == private_key
+    assert result == transaction  # Check method chaining
+
+
+def test_set_batch_key_with_public_key():
+    """Test that batch_key can be set with PublicKey."""
+    private_key = PrivateKey.generate_ed25519()
+    public_key = private_key.public_key()
+    transaction = TransferTransaction()
+    
+    result = transaction.set_batch_key(public_key)
+    
+    assert transaction.batch_key == public_key
+    assert result == transaction  # Check method chaining
+
+
+def test_batch_key_type_annotation():
+    """Test that batch_key accepts both PrivateKey and PublicKey types."""
+    transaction = TransferTransaction()
+    
+    # Test with PrivateKey
+    private_key = PrivateKey.generate_ecdsa()
+    transaction.set_batch_key(private_key)
+    assert isinstance(transaction.batch_key, PrivateKey)
+    
+    # Test with PublicKey
+    public_key = private_key.public_key()
+    transaction.set_batch_key(public_key)
+    assert isinstance(transaction.batch_key, PublicKey)
+
+
+def test_batch_key_none_by_default():
+    """Test that batch_key is None by default."""
+    transaction = TransferTransaction()
+    assert transaction.batch_key is None
